Assignment 1 - GRPC Streaming App/RedditDataApp/app/requirements.txt
grpcio
googleapis-common-protos
flask
redis
pandas